backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.10.5
python-dotenv==1.0.1